feat: reorganizing all of the metadata files
