+++ apps/spruce/src/pages/task/metadata/GeneralSection/index.tsx
@@ -1,0 +1,113 @@
+import { InlineCode } from "@leafygreen-ui/typography";
+import { Link } from "react-router-dom";
+import { StyledRouterLink } from "@evg-ui/lib/components/styles";
+import { shortenGithash } from "@evg-ui/lib/utils/string";
+import { useTaskAnalytics } from "analytics";
+import {
+  MetadataHeader,
+  MetadataItem,
+  MetadataLabel,
+} from "components/MetadataCard";
+import { getProjectPatchesRoute, getTaskRoute } from "constants/routes";
+import { TaskQuery } from "gql/generated/types";
+import { TaskOwnership } from "./TaskOwnership";
+
+type Task = NonNullable<TaskQuery["task"]>;
+
+interface GeneralSectionProps {
+  task: Task;
+}
+
+export const GeneralSection: React.FC<GeneralSectionProps> = ({ task }) => {
+  const taskAnalytics = useTaskAnalytics();
+
+  const {
+    baseTask,
+    displayTask,
+    execution,
+    generatedBy,
+    generatedByName,
+    project,
+    versionMetadata,
+  } = task;
+
+  const {
+    id: projectID,
+    identifier: projectIdentifier,
+    owner,
+    repo,
+  } = project || {};
+  const baseCommit = shortenGithash(
+    baseTask?.versionMetadata?.revision ?? undefined,
+  );
+  const { user } = versionMetadata ?? {};
+
+  return (
+    <>
+      <MetadataHeader title="General" />
+      <MetadataItem data-cy="task-metadata-project">
+        <MetadataLabel>Project:</MetadataLabel>{" "}
+        <StyledRouterLink
+          data-cy="project-link"
+          onClick={() =>
+            taskAnalytics.sendEvent({
+              name: "Clicked metadata link",
+              "link.type": "project link",
+            })
+          }
+          to={getProjectPatchesRoute(projectIdentifier || projectID || "")}
+        >
+          {projectIdentifier || `${owner}/${repo}`}
+        </StyledRouterLink>
+      </MetadataItem>
+      <MetadataItem>
+        <MetadataLabel>Submitted by:</MetadataLabel> {user.userId}
+      </MetadataItem>
+      {generatedBy && (
+        <MetadataItem>
+          <MetadataLabel>Generated by:</MetadataLabel>{" "}
+          <StyledRouterLink to={getTaskRoute(generatedBy)}>
+            {generatedByName}
+          </StyledRouterLink>
+        </MetadataItem>
+      )}
+      <TaskOwnership execution={execution} taskId={task.id} />
+      {displayTask && (
+        <MetadataItem>
+          <MetadataLabel>Display Task:</MetadataLabel>{" "}
+          <StyledRouterLink
+            onClick={() =>
+              taskAnalytics.sendEvent({
+                name: "Clicked metadata link",
+                "link.type": "display task link",
+              })
+            }
+            to={getTaskRoute(displayTask.id, {
+              execution: displayTask.execution,
+            })}
+          >
+            {displayTask.displayName}
+          </StyledRouterLink>
+        </MetadataItem>
+      )}
+      {baseTask && (
+        <MetadataItem>
+          <MetadataLabel>Base commit:</MetadataLabel>{" "}
+          <InlineCode
+            as={Link}
+            data-cy="base-task-link"
+            onClick={() =>
+              taskAnalytics.sendEvent({
+                name: "Clicked metadata link",
+                "link.type": "base commit",
+              })
+            }
+            to={getTaskRoute(baseTask.id, { execution: baseTask.execution })}
+          >
+            {baseCommit}
+          </InlineCode>
+        </MetadataItem>
+      )}
+    </>
+  );
+};
